Add functionality to check user privilege and extract their ldap in CommentData classes
- Only people who login with @google emails can utilize the comment section now

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteCommentDataServlet.java
@@ -15,6 +15,10 @@ public class DeleteCommentDataServlet extends CommentDataServlet {
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    if (!hasEnoughPrivilege()) {
+      return;
+    }
+
     deleteAll(ENTITY_KIND);
     response.sendRedirect(REDIRECT_URL_HOME);
   }
